fixed some auth. bugs

+++ src/Controller/EnrollmentsController.php
@@ -19,59 +19,97 @@ class EnrollmentsController extends AbstractController
      */
     public function index($studentId)
     {
-        //napravit vamo kontrolu jel isti id usera i studenta kojen se list minja, ako je ROLE_USER
-        // i za assign i delete isto
-        $doctrine = $this->getDoctrine();
-        $repository = $doctrine->getRepository(StudentSubject::class);
-        $student = $doctrine->getRepository(Student::class)->find($studentId);
+        $user = $this->getDoctrine()->getRepository(Student::class)->findOneByEmail((string)($this->getUser()));
 
-        //$assignedSubjects = $student->getStudentSubjects();   //ne moze
-        $assignedSubjects = $repository->findSubjectsAssignedToStudent($student);
+        if($this->isGranted('ROLE_ADMIN') || $user->getId() == $studentId) {
+            $doctrine = $this->getDoctrine();
+            $repository = $doctrine->getRepository(StudentSubject::class);
+            $student = $doctrine->getRepository(Student::class)->find($studentId);
 
-        if($student->getStatus() == 'redovan')
-            $allSubjects = $doctrine->getRepository(Subject::class)->findAllBySemRedovni();
-        else
-            $allSubjects = $doctrine->getRepository(Subject::class)->findAllBySemIzvanredni();
+            //$assignedSubjects = $student->getStudentSubjects();   //ne moze
+            $assignedSubjects = $repository->findSubjectsAssignedToStudent($student);
+            $studentSubjects = $assignedSubjects;
 
-        $unassignedSubjects = array_diff($allSubjects, $assignedSubjects);
+            if ($student->getStatus() == 'redovan')
+                $allSubjects = $doctrine->getRepository(Subject::class)->findAllBySemRedovni();
+            else
+                $allSubjects = $doctrine->getRepository(Subject::class)->findAllBySemIzvanredni();
 
-        $assignedSubjects = array_diff($allSubjects, $unassignedSubjects);
+            $unassignedSubjects = array_diff($allSubjects, $assignedSubjects);
 
-        return $this->render("enrollments/index.html.twig", [
-            "student" => [
-                "id" => $studentId,
-                "email" => $student->getEmail(),
-                "status" => $student->getStatus(),
-            ],
-            "assignedSubjects" => $assignedSubjects
-            ,
-            "unassignedSubjects" => $unassignedSubjects
-        ]);
+            $assignedSubjects = array_diff($allSubjects, $unassignedSubjects);
+
+
+            return $this->render("enrollments/index.html.twig", [
+                "student" => [
+                    "id" => $studentId,
+                    "email" => $student->getEmail(),
+                    "status" => $student->getStatus(),
+                ],
+                "assignedSubjects" => $assignedSubjects
+                ,
+                "unassignedSubjects" => $unassignedSubjects
+                ,
+                "studentSubjects" => $studentSubjects
+            ]);
+        }
+
+        return $this->redirectToRoute('enrollments.index', ['studentId' => $user->getId()]);
     }
 
     /**
      * @Route("/{studentId}/assign/{subjectId}", name="enrollments.assign")
      */
-    public function assign($studentId, $subjectId)  //stavit da samo mentor moze stavit "passed"
+    public function assign($studentId, $subjectId)
     {
-        $doctrine = $this->getDoctrine();
+        $user = $this->getDoctrine()->getRepository(Student::class)->findOneByEmail((string)($this->getUser()));
 
-        $student = $doctrine->getRepository(Student::class)->find($studentId);
-        $subject = $doctrine->getRepository(Subject::class)->find($subjectId);
+        if($this->isGranted('ROLE_ADMIN') || $user->getId() == $studentId) {
+            $doctrine = $this->getDoctrine();
 
-        $studentSubject = new StudentSubject();
+            $student = $doctrine->getRepository(Student::class)->find($studentId);
+            $subject = $doctrine->getRepository(Subject::class)->find($subjectId);
 
-        $studentSubject
-            ->setStudent($student)
-            ->setSubject($subject)
-            ->setStatus("enrolled")
-        ;
+            $studentSubject = new StudentSubject();
 
-        $em = $doctrine->getManager();
-        $em->persist($studentSubject);
-        $em->flush();
+            $studentSubject
+                ->setStudent($student)
+                ->setSubject($subject)
+                ->setStatus("enrolled");
 
-        return $this->redirectToRoute('enrollments.index', ['studentId' => $studentId]);
+            $em = $doctrine->getManager();
+            $em->persist($studentSubject);
+            $em->flush();
+
+            return $this->redirectToRoute('enrollments.index', ['studentId' => $studentId]);
+        }
+
+        return $this->redirectToRoute('enrollments.index', ['studentId' => $user->getId()]);
+    }
+
+    /**
+     * @Route("/{studentId}/pass/{subjectId}", name="enrollments.pass")
+     */
+    public function pass($studentId, $subjectId)
+    {
+        $user = $this->getDoctrine()->getRepository(Student::class)->findOneByEmail((string)($this->getUser()));
+
+        if($this->isGranted('ROLE_ADMIN') || $user->getId() == $studentId) {
+            $doctrine = $this->getDoctrine();
+            $em = $doctrine->getManager();
+
+            $ss = $em->getRepository(StudentSubject::class)->findOneById($studentId, $subjectId);
+            if ($ss->getStatus() == 'enrolled')
+                $ss->setStatus("passed");
+            else
+                $ss->setStatus("enrolled");
+
+            $em->persist($ss);
+            $em->flush();
+
+            return $this->redirectToRoute('enrollments.index', ['studentId' => $studentId]);
+        }
+        return $this->redirectToRoute('enrollments.index', ['studentId' => $user->getId()]);
     }
 
     /**
@@ -79,19 +117,25 @@ class EnrollmentsController extends AbstractController
      */
     public function remove($studentId, $subjectId)
     {
-        $doctrine = $this->getDoctrine();
+        $user = $this->getDoctrine()->getRepository(Student::class)->findOneByEmail((string)($this->getUser()));
 
-        $student = $doctrine->getRepository(Student::class)->find($studentId);
-        $subject = $doctrine->getRepository(Subject::class)->find($subjectId);
-        $rep = $doctrine->getRepository(StudentSubject::class);
+        if($this->isGranted('ROLE_ADMIN') || $user->getId() == $studentId) {
+            $doctrine = $this->getDoctrine();
 
-        $sS = $rep->findOneBy(['student' => $student, 'subject' => $subject]);
+            $student = $doctrine->getRepository(Student::class)->find($studentId);
+            $subject = $doctrine->getRepository(Subject::class)->find($subjectId);
+            $rep = $doctrine->getRepository(StudentSubject::class);
 
-        $em = $doctrine->getManager();
-        $em->remove($sS);
-        $em->flush();
+            $sS = $rep->findOneBy(['student' => $student, 'subject' => $subject]);
 
-        return $this->redirectToRoute('enrollments.index', ['studentId' => $studentId]);
+            $em = $doctrine->getManager();
+            $em->remove($sS);
+            $em->flush();
+
+            return $this->redirectToRoute('enrollments.index', ['studentId' => $studentId]);
+        }
+
+        return $this->redirectToRoute('enrollments.index', ['studentId' => $user->getId()]);
     }
 
 }
